requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
langchain-text-splitters
langchain-openai
python-dotenv
faiss-cpu
numpy
pypdf
tiktoken